+++ src/features/TodoListItem/TodoListItem.test.jsx
@@ -1,6 +1,5 @@
 import { render, screen } from '@testing-library/react';
 import { describe, it, expect, vi } from 'vitest';
-import React from 'react';
 import userEvent from '@testing-library/user-event';
 
 import TodoListItem from './TodoListItem.component';